expRate to float

+++ src/ml/qLearningAgent.cpp
@@ -3,7 +3,7 @@
 PossibleMove QLearningAgent::chooseAction(Grid& grid) {
     std::vector<PossibleMove> possibleMoves = Game::getPossibleMoves(this->_player, grid);
 
-    if (static_cast<double>(rand()) / RAND_MAX <= this->_expRate) {
+    if (static_cast<float>(rand()) / RAND_MAX <= this->_expRate) {
 
         return possibleMoves.at(rand() % possibleMoves.size());
     } else {
